refactor(web): extract useSandboxActions hook from page.tsx
Move the sandbox/repo action handlers (env-var fetch/save, project download,
open-in-VS-Code, open-on-GitHub, and git clone/checkout clipboard helpers) plus
their local state into lib/hooks/useSandboxActions.ts.
